move test into api_buildback_test
